+++ web/app/(commonLayout)/apps/page.tsx
@@ -1,8 +1,7 @@
 import classNames from 'classnames'
 import style from '../list.module.css'
 import Apps from './Apps'
-import { getLocaleOnServer } from '@/i18n/server'
-import { useTranslation as translate } from '@/i18n/i18next-serverside-config'
+import { getLocaleOnServer, useTranslation as translate } from '@/i18n/server'
 
 const AppList = async () => {
   const locale = getLocaleOnServer()
