+++ web/app/(commonLayout)/datasets/(datasetDetailLayout)/[datasetId]/settings/page.tsx
@@ -1,15 +1,12 @@
 import React from 'react'
-import { getLocaleOnServer } from '@/i18n/server'
-import { useTranslation as translate } from '@/i18n/i18next-serverside-config'
+import { getLocaleOnServer, useTranslation as translate } from '@/i18n/server'
 import Form from '@/app/components/datasets/settings/form'
 
 type Props = {
   params: { datasetId: string }
 }
 
-const Settings = async ({
-  params: { datasetId },
-}: Props) => {
+const Settings = async () => {
   const locale = getLocaleOnServer()
   const { t } = await translate(locale, 'dataset-settings')
 
@@ -19,7 +16,7 @@ const Settings = async ({
         <div className='mb-1 text-lg font-semibold text-gray-900'>{t('title')}</div>
         <div className='text-sm text-gray-500'>{t('desc')}</div>
       </div>
-      <Form datasetId={datasetId} />
+      <Form />
     </div>
   )
 }
